refactor(test): P5-1 extract pulp #1434 per-edge margin/padding tests
Extracts the cross-surface mega-batch per-edge margin/padding cluster
from test_widget_bridge.cpp into test_widget_bridge_css_per_edge.cpp
(17 TEST_CASEs, 73 assertions, 573 lines). Pins the bridge ↔ Yoga node
↔ ParsedStyle plumbing for marginTop / marginRight / marginBottom /
marginLeft (and padding counterparts) — each side maps to its own
Yoga edge enum without cross-contamination.

test_widget_bridge.cpp: 5,483 → 4,945 (-538 lines)
