+++ test/test_widget_bridge.cpp
@@ -3054,544 +3054,6 @@ TEST_CASE("setTextAlign on a container View accepts match-parent (encoded as 5)"
     REQUIRE(inh.value() == 5);
 }
 
-// ── pulp #1434 (cross-surface mega-batch) — per-edge margin/padding
-// accept percent strings + auto (margin only). Mirrors width/height
-// (#1426) and top/right/bottom/left (#1451) percent patterns. Yoga
-// dispatches `dim_*.unit == percent` to `YGNodeStyleSetMargin/PaddingPercent`;
-// `unit == auto_` (margin only) to `YGNodeStyleSetMarginAuto`.
-
-TEST_CASE("setFlex padding edges accept percent strings",
-          "[view][bridge][css][issue-1434-edges]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setFlex('a', 'padding_top',    '10%');
-        setFlex('a', 'padding_right',  '20%');
-        setFlex('a', 'padding_bottom', '30%');
-        setFlex('a', 'padding_left',   '40%');
-    )");
-
-    const auto& f = bridge.widget("a")->flex();
-    REQUIRE(f.dim_padding_top.unit    == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_padding_top.value,    WithinAbs(10.0f, 0.001f));
-    REQUIRE(f.dim_padding_right.unit  == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_padding_right.value,  WithinAbs(20.0f, 0.001f));
-    REQUIRE(f.dim_padding_bottom.unit == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_padding_bottom.value, WithinAbs(30.0f, 0.001f));
-    REQUIRE(f.dim_padding_left.unit   == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_padding_left.value,   WithinAbs(40.0f, 0.001f));
-}
-
-TEST_CASE("setFlex padding edges numeric path stays px",
-          "[view][bridge][css][issue-1434-edges]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setFlex('a', 'padding_top',    8);
-        setFlex('a', 'padding_right',  12);
-        setFlex('a', 'padding_bottom', 16);
-        setFlex('a', 'padding_left',   4);
-    )");
-
-    const auto& f = bridge.widget("a")->flex();
-    REQUIRE(f.dim_padding_top.unit == DimensionUnit::px);
-    REQUIRE_THAT(f.padding_top,    WithinAbs(8.0f,  0.001f));
-    REQUIRE_THAT(f.padding_right,  WithinAbs(12.0f, 0.001f));
-    REQUIRE_THAT(f.padding_bottom, WithinAbs(16.0f, 0.001f));
-    REQUIRE_THAT(f.padding_left,   WithinAbs(4.0f,  0.001f));
-}
-
-TEST_CASE("setFlex margin edges accept percent strings",
-          "[view][bridge][css][issue-1434-edges]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setFlex('a', 'margin_top',    '5%');
-        setFlex('a', 'margin_right',  '10%');
-        setFlex('a', 'margin_bottom', '15%');
-        setFlex('a', 'margin_left',   '20%');
-    )");
-
-    const auto& f = bridge.widget("a")->flex();
-    REQUIRE(f.dim_margin_top.unit    == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_margin_top.value,    WithinAbs(5.0f,  0.001f));
-    REQUIRE(f.dim_margin_right.unit  == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_margin_right.value,  WithinAbs(10.0f, 0.001f));
-    REQUIRE(f.dim_margin_bottom.unit == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_margin_bottom.value, WithinAbs(15.0f, 0.001f));
-    REQUIRE(f.dim_margin_left.unit   == DimensionUnit::percent);
-    REQUIRE_THAT(f.dim_margin_left.value,   WithinAbs(20.0f, 0.001f));
-}
-
-TEST_CASE("setFlex margin edges accept 'auto' keyword",
-          "[view][bridge][css][issue-1434-edges]") {
-    // `marginLeft: 'auto'; marginRight: 'auto'` is the canonical
-    // centering idiom in CSS / RN; Yoga supports it via
-    // YGNodeStyleSetMarginAuto.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setFlex('a', 'margin_left',  'auto');
-        setFlex('a', 'margin_right', 'auto');
-    )");
-
-    const auto& f = bridge.widget("a")->flex();
-    REQUIRE(f.dim_margin_left.unit  == DimensionUnit::auto_);
-    REQUIRE(f.dim_margin_right.unit == DimensionUnit::auto_);
-}
-
-// ── pulp DIVERGE→PASS sweep — canvas2d surface ──────────────────────────
-
-TEST_CASE("ctx.fill('evenodd') threads fillRule int_val through bridge",
-          "[view][bridge][canvas][diverge-pass-canvas2d]") {
-    // pulp DIVERGE→PASS sweep — `ctx.fill('evenodd')` and
-    // `ctx.clip('evenodd')` now thread the fillRule keyword through
-    // to the bridge as int_val (0=nonzero, 1=evenodd). Pairs with
-    // [issue-1522] which tests the bridge fns directly.
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    root.set_theme(Theme::dark());
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var c = document.createElement('canvas');
-        c.id = 'fr-shim-canvas';
-        c.width = 100; c.height = 100;
-        document.body.appendChild(c);
-        var ctx = c.getContext('2d');
-        ctx.beginPath();
-        ctx.moveTo(0, 0); ctx.lineTo(10, 0);
-        ctx.lineTo(10, 10); ctx.lineTo(0, 10);
-        ctx.closePath();
-        ctx.fill('evenodd');
-        ctx.beginPath();
-        ctx.moveTo(0, 0); ctx.lineTo(10, 0);
-        ctx.lineTo(10, 10); ctx.closePath();
-        ctx.fill();             // default nonzero
-        ctx.beginPath();
-        ctx.moveTo(0, 0); ctx.lineTo(10, 0);
-        ctx.lineTo(10, 10); ctx.closePath();
-        ctx.clip('evenodd');
-        ctx.beginPath();
-        ctx.moveTo(0, 0); ctx.lineTo(10, 0);
-        ctx.lineTo(10, 10); ctx.closePath();
-        ctx.clip();
-    )");
-
-    auto* canvas = canvasFromBridge(bridge, engine, "fr-shim-canvas");
-    REQUIRE(canvas != nullptr);
-    using T = pulp::view::CanvasDrawCmd::Type;
-    std::vector<int> fill_int_vals;
-    std::vector<int> clip_int_vals;
-    for (const auto& cmd : canvas->commands()) {
-        if (cmd.type == T::fill_path) fill_int_vals.push_back(cmd.int_val);
-        if (cmd.type == T::clip)      clip_int_vals.push_back(cmd.int_val);
-    }
-    REQUIRE(fill_int_vals.size() == 2);
-    REQUIRE(fill_int_vals[0] == 1);  // explicit evenodd
-    REQUIRE(fill_int_vals[1] == 0);  // default nonzero
-    REQUIRE(clip_int_vals.size() == 2);
-    REQUIRE(clip_int_vals[0] == 1);  // explicit evenodd
-    REQUIRE(clip_int_vals[1] == 0);  // default nonzero
-}
-
-// ── pulp DIVERGE→PASS sweep — html surface ──────────────────────────────
-
-TEST_CASE("Element.disabled wires to setEnabled",
-          "[view][bridge][html][diverge-pass-html]") {
-    // Previously `el.disabled = true` only flipped the stylesheet flag;
-    // the underlying widget kept handling pointer events. Now wired
-    // end-to-end via setEnabled. Drive via JS so the assertion runs
-    // against the real CSSStyleDeclaration / disabled property accessor.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var input = document.createElement('input');
-        input.id = 'box';
-        input.type = 'text';
-        document.body.appendChild(input);
-        globalThis.__test_input_id__ = input._id;
-        input.disabled = true;
-        globalThis.__test_disabled_value__ = input.disabled;
-    )");
-
-    auto id = engine.evaluate("globalThis.__test_input_id__").getWithDefault<std::string>("");
-    REQUIRE_FALSE(id.empty());
-    auto* w = bridge.widget(id);
-    REQUIRE(w != nullptr);
-    REQUIRE_FALSE(w->enabled());
-    auto js_disabled = engine.evaluate("globalThis.__test_disabled_value__").getWithDefault<bool>(false);
-    REQUIRE(js_disabled);
-}
-
-TEST_CASE("new Event() round-trips through dispatchEvent",
-          "[view][bridge][html][diverge-pass-html]") {
-    // The Event / CustomEvent constructors live in web-compat-element.js.
-    // Verify `new Event('foo')` produces a target-able event that
-    // dispatchEvent fires through addEventListener listeners.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var el = document.createElement('div');
-        document.body.appendChild(el);
-        globalThis.__test_event_count__ = 0;
-        globalThis.__test_event_type__ = '';
-        globalThis.__test_event_detail__ = null;
-        el.addEventListener('foo', function(e) {
-            globalThis.__test_event_count__++;
-            globalThis.__test_event_type__ = e.type;
-        });
-        el.addEventListener('bar', function(e) {
-            globalThis.__test_event_detail__ = e.detail;
-        });
-        el.dispatchEvent(new Event('foo'));
-        el.dispatchEvent(new CustomEvent('bar', { detail: 42 }));
-    )");
-
-    auto count  = engine.evaluate("globalThis.__test_event_count__").getWithDefault<int>(-1);
-    auto type   = engine.evaluate("globalThis.__test_event_type__").getWithDefault<std::string>("");
-    auto detail = engine.evaluate("globalThis.__test_event_detail__").getWithDefault<int>(-1);
-    REQUIRE(count == 1);
-    REQUIRE(type == "foo");
-    REQUIRE(detail == 42);
-}
-
-TEST_CASE("<dialog> show/close round-trips visibility and dispatches close",
-          "[view][bridge][html][diverge-pass-html]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var d = document.createElement('dialog');
-        document.body.appendChild(d);
-        globalThis.__test_dialog_id__ = d._id;
-        globalThis.__test_dialog_close_count__ = 0;
-        d.addEventListener('close', function() {
-            globalThis.__test_dialog_close_count__++;
-        });
-        // Initially hidden.
-        globalThis.__test_dialog_open_initial__ = d.open;
-        d.show();
-        globalThis.__test_dialog_open_after_show__ = d.open;
-        d.close('ok');
-        globalThis.__test_dialog_open_after_close__ = d.open;
-        globalThis.__test_dialog_return_value__ = d.returnValue;
-    )");
-
-    auto initial      = engine.evaluate("globalThis.__test_dialog_open_initial__").getWithDefault<bool>(true);
-    auto after_show   = engine.evaluate("globalThis.__test_dialog_open_after_show__").getWithDefault<bool>(false);
-    auto after_close  = engine.evaluate("globalThis.__test_dialog_open_after_close__").getWithDefault<bool>(true);
-    auto close_count  = engine.evaluate("globalThis.__test_dialog_close_count__").getWithDefault<int>(-1);
-    auto return_value = engine.evaluate("globalThis.__test_dialog_return_value__").getWithDefault<std::string>("");
-    REQUIRE_FALSE(initial);
-    REQUIRE(after_show);
-    REQUIRE_FALSE(after_close);
-    REQUIRE(close_count == 1);
-    REQUIRE(return_value == "ok");
-}
-
-TEST_CASE("<details> open setter dispatches toggle event",
-          "[view][bridge][html][diverge-pass-html]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var d = document.createElement('details');
-        document.body.appendChild(d);
-        globalThis.__test_toggle_count__ = 0;
-        d.addEventListener('toggle', function() {
-            globalThis.__test_toggle_count__++;
-        });
-        globalThis.__test_open_initial__ = d.open;
-        d.open = true;
-        globalThis.__test_open_after_set__ = d.open;
-        globalThis.__test_open_attribute__ = d.getAttribute('open');
-        d.open = false;
-        globalThis.__test_open_after_unset__ = d.open;
-    )");
-
-    auto initial    = engine.evaluate("globalThis.__test_open_initial__").getWithDefault<bool>(true);
-    auto after_set  = engine.evaluate("globalThis.__test_open_after_set__").getWithDefault<bool>(false);
-    auto attr       = engine.evaluate("globalThis.__test_open_attribute__").getWithDefault<std::string>("__missing__");
-    auto after_unset = engine.evaluate("globalThis.__test_open_after_unset__").getWithDefault<bool>(true);
-    auto toggles    = engine.evaluate("globalThis.__test_toggle_count__").getWithDefault<int>(-1);
-    REQUIRE_FALSE(initial);
-    REQUIRE(after_set);
-    REQUIRE(attr == "");  // present, value empty
-    REQUIRE_FALSE(after_unset);
-    REQUIRE(toggles == 2);
-}
-
-TEST_CASE("<label for=...> click toggles labeled checkbox",
-          "[view][bridge][html][diverge-pass-html]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var cb = document.createElement('input');
-        cb.type = 'checkbox';
-        cb.id = 'mybox';
-        document.body.appendChild(cb);
-        var lbl = document.createElement('label');
-        lbl.setAttribute('for', 'mybox');
-        document.body.appendChild(lbl);
-
-        globalThis.__test_input_event_count__ = 0;
-        cb.addEventListener('input', function() {
-            globalThis.__test_input_event_count__++;
-        });
-        // Synthesize a click on the label.
-        var evt = new Event('click', { bubbles: true });
-        lbl.dispatchEvent(evt);
-        globalThis.__test_checked_after_click__ = cb.checked;
-    )");
-
-    auto count   = engine.evaluate("globalThis.__test_input_event_count__").getWithDefault<int>(-1);
-    auto checked = engine.evaluate("globalThis.__test_checked_after_click__").getWithDefault<bool>(false);
-    REQUIRE(count == 1);
-    REQUIRE(checked);
-}
-
-TEST_CASE("addEventListener('wheel', fn) routes through registerWheel",
-          "[view][bridge][html][diverge-pass-html]") {
-    // Verify the wheel branch in _registerNativeEvent is exercised — we
-    // can't easily generate a wheel scroll from C++ without driving the
-    // event system, so smoke-test that the call doesn't throw and the
-    // listener is recorded.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var el = document.createElement('div');
-        document.body.appendChild(el);
-        var fired = false;
-        el.addEventListener('wheel', function(e) {
-            fired = true;
-        });
-        // Manually invoke the bridge's __dispatch__ for 'wheel' to
-        // simulate a native wheel event delivery — that's the path
-        // _registerNativeEvent wires up via on(id, 'wheel', ...).
-        if (typeof __dispatch__ === 'function') {
-            __dispatch__(el._id, 'wheel', 5, -7);
-        }
-        globalThis.__test_wheel_fired__ = fired;
-    )");
-
-    auto fired = engine.evaluate("globalThis.__test_wheel_fired__").getWithDefault<bool>(false);
-    REQUIRE(fired);
-}
-
-TEST_CASE("addEventListener('drop', fn) routes through registerDrop",
-          "[view][bridge][html][diverge-pass-html]") {
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var el = document.createElement('div');
-        document.body.appendChild(el);
-        globalThis.__test_drop_payload__ = '';
-        el.addEventListener('drop', function(e) {
-            globalThis.__test_drop_payload__ = e._dropData
-                ? (e._dropData.type + ':' + e._dropData.data)
-                : 'no-data';
-        });
-        // The drop handler is registered as a global function named
-        // __drop_cb_<id>; invoke it directly to simulate the native
-        // drop completion path.
-        var cbName = '__drop_cb_' + el._id.replace(/[^a-zA-Z0-9_]/g, '_');
-        if (typeof globalThis[cbName] === 'function') {
-            globalThis[cbName]('text', 'hello world', 10, 20);
-        }
-    )");
-
-    auto payload = engine.evaluate("globalThis.__test_drop_payload__").getWithDefault<std::string>("");
-    REQUIRE(payload == "text:hello world");
-}
-
-TEST_CASE("setFlex start/end accept 'auto' keyword",
-          "[view][bridge][css][diverge-pass-yoga]") {
-    // pulp DIVERGE→PASS sweep — yoga/start and yoga/end claimed `auto`
-    // as unsupported. Yoga DOES expose YGNodeStyleSetPositionAuto, so
-    // wire `'auto'` through the bridge → FlexStyle::dim_start/dim_end
-    // → yoga_layout.cpp::apply_logical_position. The keyword routes via
-    // DimensionUnit::auto_.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setFlex('a', 'start', 'auto');
-        setFlex('a', 'end',   'auto');
-    )");
-
-    const auto& f = bridge.widget("a")->flex();
-    REQUIRE(f.dim_start.unit == DimensionUnit::auto_);
-    REQUIRE(f.dim_end.unit   == DimensionUnit::auto_);
-}
-
-TEST_CASE("flex shorthand keyword forms",
-          "[view][bridge][css][diverge-pass-yoga]") {
-    // pulp DIVERGE→PASS sweep — yoga/flex claimed `'auto'`/`'none'`/
-    // `'initial'` as unsupported because the JS shorthand parser
-    // parseFloat()ed the keyword to NaN. The CSS Flexible Box spec
-    // expansion is:
-    //   flex: auto    ≡ 1 1 auto
-    //   flex: none    ≡ 0 0 auto
-    //   flex: initial ≡ 0 1 auto
-    // Drives the CSSStyleDeclaration shim (the actual code path that
-    // contains the new keyword expansion) and verifies each keyword
-    // reaches FlexStyle with the correct grow / shrink / basis-unit.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        createPanel('c', '');
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        var sc = new CSSStyleDeclaration({ _id: 'c', _nativeCreated: true });
-        sa.flex = 'auto';
-        sb.flex = 'none';
-        sc.flex = 'initial';
-    )");
-
-    const auto& fa = bridge.widget("a")->flex();
-    REQUIRE_THAT(fa.flex_grow,   WithinAbs(1.0, 0.0001));
-    REQUIRE_THAT(fa.flex_shrink, WithinAbs(1.0, 0.0001));
-    REQUIRE(fa.dim_flex_basis.unit == DimensionUnit::auto_);
-
-    const auto& fb = bridge.widget("b")->flex();
-    REQUIRE_THAT(fb.flex_grow,   WithinAbs(0.0, 0.0001));
-    REQUIRE_THAT(fb.flex_shrink, WithinAbs(0.0, 0.0001));
-    REQUIRE(fb.dim_flex_basis.unit == DimensionUnit::auto_);
-
-    const auto& fc = bridge.widget("c")->flex();
-    REQUIRE_THAT(fc.flex_grow,   WithinAbs(0.0, 0.0001));
-    REQUIRE_THAT(fc.flex_shrink, WithinAbs(1.0, 0.0001));
-    REQUIRE(fc.dim_flex_basis.unit == DimensionUnit::auto_);
-}
-
-TEST_CASE("setOverflow accepts scroll keyword",
-          "[view][bridge][diverge-pass-yoga]") {
-    // pulp DIVERGE→PASS sweep — yoga/overflow claimed `scroll` as
-    // unsupported. View::Overflow now has 3 values; the bridge accepts
-    // 'scroll' as a third keyword and yoga_layout.cpp forwards it
-    // through YGNodeStyleSetOverflow. Paint clipping treats scroll
-    // like hidden (no scrollbar UI yet), but the keyword survives the
-    // round-trip — closing the harness gap.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        setOverflow('a', 'scroll');
-        createPanel('b', '');
-        setOverflow('b', 'visible');
-        createPanel('c', '');
-        setOverflow('c', 'hidden');
-    )");
-
-    REQUIRE(bridge.widget("a")->overflow() == View::Overflow::scroll);
-    REQUIRE(bridge.widget("b")->overflow() == View::Overflow::visible);
-    REQUIRE(bridge.widget("c")->overflow() == View::Overflow::hidden);
-}
-
-TEST_CASE("padding_left percent caps the layout edge",
-          "[view][bridge][css][issue-1434-edges]") {
-    // End-to-end: percent reaches Yoga and produces the expected
-    // resolved pixel size after layout.
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 200});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('parent', '');
-        setFlex('parent', 'padding_left', '25%');
-        setFlex('parent', 'width',  '100%');
-        setFlex('parent', 'height', '100%');
-    )");
-
-    auto* parent = bridge.widget("parent");
-    REQUIRE(parent != nullptr);
-    root.layout_children();
-    // 25% of 400px parent width → 100px padding-left.
-    REQUIRE_THAT(parent->bounds().width, WithinAbs(400.0f, 0.5f));
-}
-
-TEST_CASE("CSSStyleDeclaration forwards marginTop percent + auto verbatim",
-          "[view][bridge][css][issue-1434-edges]") {
-    // The DOM-lite el.style adapter must forward 'NN%' and 'auto' as
-    // strings so the bridge's per-unit branch is reached.
-    ScriptEngine engine;
-    View root;
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createPanel('a', '');
-        createPanel('b', '');
-        var sa = new CSSStyleDeclaration({ _id: 'a', _nativeCreated: true });
-        var sb = new CSSStyleDeclaration({ _id: 'b', _nativeCreated: true });
-        sa._applyProperty('marginTop',  '50%');
-        sa._applyProperty('paddingTop', '25%');
-        sb._applyProperty('marginLeft',  'auto');
-        sb._applyProperty('marginRight', 'auto');
-    )");
-
-    const auto& fa = bridge.widget("a")->flex();
-    REQUIRE(fa.dim_margin_top.unit    == DimensionUnit::percent);
-    REQUIRE_THAT(fa.dim_margin_top.value,    WithinAbs(50.0f, 0.001f));
-    REQUIRE(fa.dim_padding_top.unit   == DimensionUnit::percent);
-    REQUIRE_THAT(fa.dim_padding_top.value,   WithinAbs(25.0f, 0.001f));
-
-    const auto& fb = bridge.widget("b")->flex();
-    REQUIRE(fb.dim_margin_left.unit  == DimensionUnit::auto_);
-    REQUIRE(fb.dim_margin_right.unit == DimensionUnit::auto_);
-}
-
 // ── pulp #1434 small-wins bundle (Triage #7 + #14) ──────────────────────
 //
 // Triage #7: cursor enum fan-out — extended setCursor case ladder maps
@@ -4980,504 +4442,3 @@ TEST_CASE("setBorderTopColor preserves uniform borderWidth shorthand",
     REQUIRE_THAT(child->bounds().height, WithinAbs(80.0f, 0.5f));
 }
 
-// ── pulp #1737 RN-OOS-fixup (catalog audit 2026-05-11) ──────────────────
-// Followup wave: 4 RN box-shadow longhand setters + 2 CSS scroll-behavior
-// slots. All 6 cited in compat.json mapsTo claims — these tests pin the
-// bridge fn surface so the audit's catalog claims are evidence-backed.
-
-TEST_CASE("WidgetBridge setShadowColor mutates View::shadow_.color + activates has_shadow_",
-          "[view][bridge][issue-1737]") {
-    using Catch::Matchers::WithinAbs;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setShadowColor('gain', '#ff0000')");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().color.r, WithinAbs(1.0f, 0.01f));
-    REQUIRE_THAT(w->box_shadow().color.g, WithinAbs(0.0f, 0.01f));
-    REQUIRE_THAT(w->box_shadow().color.b, WithinAbs(0.0f, 0.01f));
-}
-
-TEST_CASE("WidgetBridge setShadowOffset mutates offset_x / offset_y in isolation",
-          "[view][bridge][issue-1737]") {
-    using Catch::Matchers::WithinAbs;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setShadowOffset('gain', 7, 11)");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().offset_x, WithinAbs(7.0f, 0.001f));
-    REQUIRE_THAT(w->box_shadow().offset_y, WithinAbs(11.0f, 0.001f));
-}
-
-TEST_CASE("WidgetBridge setShadowOpacity writes color alpha (0..1)",
-          "[view][bridge][issue-1737]") {
-    using Catch::Matchers::WithinAbs;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setShadowOpacity('gain', 0.5)");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().color.a, WithinAbs(0.5f, 0.001f));
-}
-
-TEST_CASE("WidgetBridge setShadowRadius writes the blur field of View::shadow_",
-          "[view][bridge][issue-1737]") {
-    using Catch::Matchers::WithinAbs;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setShadowRadius('gain', 22)");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().blur, WithinAbs(22.0f, 0.001f));
-}
-
-TEST_CASE("WidgetBridge setScrollBehavior stores the keyword on View",
-          "[view][bridge][issue-1737]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setScrollBehavior('gain', 'smooth')");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->scroll_behavior() == "smooth");
-
-    bridge.load_script("setScrollBehavior('gain', 'auto')");
-    REQUIRE(w->scroll_behavior() == "auto");
-}
-
-TEST_CASE("WidgetBridge setOverscrollBehavior stores the keyword on View",
-          "[view][bridge][issue-1737]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    bridge.load_script("setOverscrollBehavior('gain', 'contain')");
-
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->overscroll_behavior() == "contain");
-
-    bridge.load_script("setOverscrollBehavior('gain', 'none')");
-    REQUIRE(w->overscroll_behavior() == "none");
-}
-
-// pulp #1737 RN-OOS-fixup final sweep — rn/elevation Material shim.
-TEST_CASE("WidgetBridge setElevation shims to Material-approx box-shadow",
-          "[view][bridge][issue-1737]") {
-    using Catch::Matchers::WithinAbs;
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-
-    // elevation 4 -> offset_y=2, blur=5, alpha≈0.19 (per the formula in
-    // widget_bridge.cpp: offset_y=max(1, n/2), blur=n+1, alpha=clamp(0.15+n*0.01, 0.15, 0.30)).
-    bridge.load_script("setElevation('gain', 4)");
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().offset_x, WithinAbs(0.0f, 0.001f));
-    REQUIRE_THAT(w->box_shadow().offset_y, WithinAbs(2.0f, 0.001f));
-    REQUIRE_THAT(w->box_shadow().blur,     WithinAbs(5.0f, 0.001f));
-    REQUIRE_THAT(w->box_shadow().color.a,  WithinAbs(0.19f, 0.01f));
-
-    // elevation 0 -> clears the shadow entirely.
-    bridge.load_script("setElevation('gain', 0)");
-    REQUIRE_FALSE(w->has_box_shadow());
-
-    // elevation 24 (max) -> alpha saturates at 0.30.
-    bridge.load_script("setElevation('gain', 24)");
-    REQUIRE(w->has_box_shadow());
-    REQUIRE_THAT(w->box_shadow().color.a, WithinAbs(0.30f, 0.001f));
-}
-
-// pulp #1737 RN-OOS-fixup (final sweep) — includeFontPadding round-trip.
-TEST_CASE("WidgetBridge setIncludeFontPadding stores the keyword on View (round-trip only)",
-          "[view][bridge][issue-1737]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-
-    // Default state — Pulp's View has include_font_padding_ = true.
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->include_font_padding() == true);
-
-    // Setting `false` (the common case — remove Android padding):
-    // Pulp accepts the keyword and stores it. Text shaping is
-    // unchanged because Pulp never had Android-vestigial padding.
-    bridge.load_script("setIncludeFontPadding('gain', false)");
-    REQUIRE(w->include_font_padding() == false);
-
-    // Setting `true` round-trips even though Pulp can't add Android-
-    // style padding — author can still query the slot.
-    bridge.load_script("setIncludeFontPadding('gain', true)");
-    REQUIRE(w->include_font_padding() == true);
-}
-
-// pulp #1737 RN-OOS-fixup #1812 — borderCurve squircle paint dispatch.
-TEST_CASE("WidgetBridge setBorderCurve toggles between circular and continuous",
-          "[view][bridge][issue-1737]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-    REQUIRE(w->border_curve() == View::BorderCurve::circular);  // default
-
-    bridge.load_script("setBorderCurve('gain', 'continuous')");
-    REQUIRE(w->border_curve() == View::BorderCurve::continuous);
-
-    bridge.load_script("setBorderCurve('gain', 'circular')");
-    REQUIRE(w->border_curve() == View::BorderCurve::circular);
-
-    // Unknown keyword falls back to circular (matches RN spec: unknown → default).
-    bridge.load_script("setBorderCurve('gain', 'banana')");
-    REQUIRE(w->border_curve() == View::BorderCurve::circular);
-}
-
-// pulp #1737 RN-OOS-fixup (final round) — isolation honest CSS-subset.
-TEST_CASE("WidgetBridge setIsolation round-trips on View slot (no paint impact)",
-          "[view][bridge][issue-1737]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script("createKnob('gain', 10, 10, 48, 48)");
-    auto* w = bridge.widget("gain");
-    REQUIRE(w != nullptr);
-
-    bridge.load_script("setIsolation('gain', 'isolate')");
-    REQUIRE(w->isolation() == "isolate");
-
-    bridge.load_script("setIsolation('gain', 'auto')");
-    REQUIRE(w->isolation() == "auto");
-}
-
-
-
-// pulp-internal Tier-1 closure for css/textTransform (2026-05-12).
-// The setTextTransform bridge already accepts the 4 CSS spec values
-// (uppercase / lowercase / capitalize / none) and routes them onto
-// Label::TextTransform. The existing widget-bridge sanity test
-// (line ~875) only exercised `uppercase` once. This focused test
-// pins the full supported value set so the row's closure (move
-// `full-width` / `full-size-kana` to arch-deferred-CJK-Unicode-width
-// in compat.json) is honest.
-TEST_CASE("setTextTransform pins all 4 CSS spec values to Label::TextTransform enum",
-          "[view][bridge][css][tier1-closure][css-textTransform]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        createLabel('upper',      'hello', '');
-        createLabel('lower',      'HELLO', '');
-        createLabel('capitalize', 'hello world', '');
-        createLabel('none_',      'hello', '');
-
-        setTextTransform('upper',      'uppercase');
-        setTextTransform('lower',      'lowercase');
-        setTextTransform('capitalize', 'capitalize');
-        setTextTransform('none_',      'none');
-    )");
-
-    auto tt = [&](const std::string& id) -> Label::TextTransform {
-        return dynamic_cast<Label*>(bridge.widget(id))->text_transform();
-    };
-    REQUIRE(tt("upper")      == Label::TextTransform::uppercase);
-    REQUIRE(tt("lower")      == Label::TextTransform::lowercase);
-    REQUIRE(tt("capitalize") == Label::TextTransform::capitalize);
-    REQUIRE(tt("none_")      == Label::TextTransform::none);
-}
-
-// pulp #1923 — drag-style interactions (FilterBank band drawing, slider
-// thumb drag, scroll gestures) lost state between pointerdown and the
-// immediately-following pointermove because safe_dispatch_eval() didn't
-// pump microtasks after dispatching the JS handler. React's setState
-// commit is queued as a microtask; without an explicit pump_message_loop
-// after engine.evaluate(), the move handler runs against the pre-down
-// state and silently bails. This test pins the contract: any side-effect
-// the pointerdown handler queues via queueMicrotask / Promise.then
-// (which is what React 18's setState uses under the hood) MUST be
-// visible to the next dispatched event on the same widget.
-TEST_CASE("WidgetBridge pumps microtasks after JS dispatch so drag-state commits before next event",
-          "[view][bridge][events][issue-1923]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    // Pointerdown queues a microtask that mutates `drag_active` (stand-in
-    // for a React setState commit). Pointermove reads `drag_active` and
-    // records what it saw at the time it fired. Pre-fix the recorded
-    // value is `false` (microtask never drained); post-fix it is `true`.
-    bridge.load_script(R"(
-        var drag_active = false;
-        var move_saw_drag_active = null;
-        var move_count = 0;
-
-        createLabel('surface', 'Surface', '');
-        on('surface', 'pointerdown', function(e) {
-            // Stand-in for React's setState commit: queue a microtask
-            // that flips the committed state. React 18's scheduler
-            // uses queueMicrotask / Promise.resolve().then for the
-            // discrete-event lane, so this is the realistic shape.
-            queueMicrotask(function() { drag_active = true; });
-        });
-        on('surface', 'pointermove', function(e) {
-            if (move_saw_drag_active === null) {
-                move_saw_drag_active = drag_active;
-            }
-            move_count++;
-        });
-        registerPointer('surface');
-    )");
-
-    auto* surface = bridge.widget("surface");
-    REQUIRE(surface != nullptr);
-    REQUIRE(surface->on_pointer_event);
-    REQUIRE(surface->on_drag);
-
-    // Sanity: nothing has run yet.
-    REQUIRE_FALSE(engine.evaluate("drag_active").getWithDefault<bool>(true));
-
-    // Dispatch pointerdown. After safe_dispatch_eval returns, the fix
-    // guarantees the queued microtask has drained and `drag_active` is
-    // committed. Pre-fix this assertion fails — the microtask is still
-    // pending in QuickJS' job queue.
-    MouseEvent down{};
-    down.is_down = true;
-    down.position = {10.0f, 10.0f};
-    down.window_position = {110.0f, 110.0f};
-    down.pointer_id = 1;
-    down.pointer_type = PointerType::mouse;
-    down.button = MouseButton::left;
-    surface->on_mouse_event(down);
-
-    REQUIRE(engine.evaluate("drag_active").getWithDefault<bool>(false));
-
-    // Now dispatch pointermove. The handler must observe the committed
-    // drag_active value (true), not the pre-down value (false). This is
-    // the user-visible regression in #1923: pointermove handlers see
-    // stale state and bail.
-    surface->on_drag({12.0f, 14.0f});
-
-    REQUIRE(engine.evaluate("move_count").getWithDefault<int>(0) == 1);
-    REQUIRE(engine.evaluate("move_saw_drag_active").getWithDefault<bool>(false));
-}
-
-// pulp #1576 — bridge-level pin for min/max width/height with %, auto,
-// and calc-family inputs. The dispatcher at web-compat-style-decl.js
-// cases minWidth / minHeight / maxWidth / maxHeight previously had a
-// dual-path (calc-family detection guard + parseCSSLength fallback);
-// #1576 collapsed both into a single resolveCSSLength call. These
-// tests pin the unified shape end-to-end through the bridge so a
-// regression that re-introduces the dual-path (or drops % preservation
-// for calc-family inputs) fails first.
-TEST_CASE("minWidth / minHeight / maxWidth / maxHeight route % and calc-family through unified resolveCSSLength",
-          "[view][bridge][css][issue-1576]") {
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        // Plain % path — unified resolveCSSLength must preserve unit='%'.
-        createPanel('mw-pct', '');
-        var sd1 = new CSSStyleDeclaration({ _id: 'mw-pct', _nativeCreated: true });
-        sd1._applyProperty('minWidth', '25%');
-
-        createPanel('mh-pct', '');
-        var sd2 = new CSSStyleDeclaration({ _id: 'mh-pct', _nativeCreated: true });
-        sd2._applyProperty('minHeight', '33%');
-
-        // calc-family % path — Codex P2: calc(50%) must reach the
-        // bridge as a percent, not absolute px.
-        createPanel('xw-calc-pct', '');
-        var sd3 = new CSSStyleDeclaration({ _id: 'xw-calc-pct', _nativeCreated: true });
-        sd3._applyProperty('maxWidth', 'min(50%, 75%)');
-
-        // calc-family px path — calc(100px + 50px) must resolve to 150px.
-        createPanel('xh-calc-px', '');
-        var sd4 = new CSSStyleDeclaration({ _id: 'xh-calc-px', _nativeCreated: true });
-        sd4._applyProperty('maxHeight', 'calc(100px + 50px)');
-    )");
-
-    auto* mw_pct      = dynamic_cast<Panel*>(bridge.widget("mw-pct"));
-    auto* mh_pct      = dynamic_cast<Panel*>(bridge.widget("mh-pct"));
-    auto* xw_calc_pct = dynamic_cast<Panel*>(bridge.widget("xw-calc-pct"));
-    auto* xh_calc_px  = dynamic_cast<Panel*>(bridge.widget("xh-calc-px"));
-    REQUIRE(mw_pct      != nullptr);
-    REQUIRE(mh_pct      != nullptr);
-    REQUIRE(xw_calc_pct != nullptr);
-    REQUIRE(xh_calc_px  != nullptr);
-
-    // % survives: dim_<dim>.unit is the percent sentinel.
-    REQUIRE(mw_pct->flex().dim_min_width.unit  == DimensionUnit::percent);
-    REQUIRE_THAT(mw_pct->flex().dim_min_width.value, WithinAbs(25.0f, 0.001f));
-    REQUIRE(mh_pct->flex().dim_min_height.unit == DimensionUnit::percent);
-    REQUIRE_THAT(mh_pct->flex().dim_min_height.value, WithinAbs(33.0f, 0.001f));
-
-    // calc-family % survives — Codex P2 pin (min(50%, 75%) = 50%).
-    REQUIRE(xw_calc_pct->flex().dim_max_width.unit  == DimensionUnit::percent);
-    REQUIRE_THAT(xw_calc_pct->flex().dim_max_width.value, WithinAbs(50.0f, 0.001f));
-
-    // calc-family px resolves to 150.
-    REQUIRE(xh_calc_px->flex().dim_max_height.unit == DimensionUnit::px);
-    REQUIRE_THAT(xh_calc_px->flex().dim_max_height.value, WithinAbs(150.0f, 0.001f));
-}
-
-// Focus-guard for global key shortcuts. Bare-key shortcuts (no Ctrl/Alt/
-// Meta/Cmd; Shift alone counts as bare since it just picks the upper-case
-// glyph) must NOT fire while a text input has focus — otherwise typing a
-// `?` into a search box would open the global cheatsheet. Modifier chords
-// (Cmd+S, Cmd+,) are always-global by design and must still fire when an
-// input is focused.
-//
-// Prereq for the default-shortcuts pass (planning/2026-05-16-default-
-// keyboard-shortcuts.md), which adds a bare-`?` cheatsheet binding.
-TEST_CASE("WidgetBridge focus-guard: bare-key shortcuts suppressed while text input focused",
-          "[view][bridge][shortcuts][focus-guard]") {
-    using namespace pulp::view;
-
-    // Always start the test with a clean focus slot so prior cases don't
-    // bleed in (the slot is a static and tests share the same process).
-    View::focused_input_ = nullptr;
-
-    ScriptEngine engine;
-    View root;
-    root.set_bounds({0, 0, 400, 300});
-    StateStore store;
-    WidgetBridge bridge(engine, root, store);
-
-    bridge.load_script(R"(
-        var bare_count = 0;
-        var shift_count = 0;
-        var cmd_count = 0;
-        function bareHit()  { bare_count++; }
-        function shiftHit() { shift_count++; }
-        function cmdHit()   { cmd_count++; }
-
-        // Bare-`?` (no modifiers): cheatsheet pattern.
-        registerShortcut(63, 0, 'bareHit');
-        // Shift+`?` (kModShift = 1): also bare for guard purposes — Shift
-        // alone is just the glyph selector when typing.
-        registerShortcut(63, 1, 'shiftHit');
-        // Cmd+`,` (kModCmd = 16): always-global modifier chord.
-        registerShortcut(44, 16, 'cmdHit');
-    )");
-
-    SECTION("no input focused: bare-key fires") {
-        REQUIRE(View::focused_input_ == nullptr);
-        bridge.forward_key_event(63, 0, true);
-        REQUIRE(engine.evaluate("bare_count").getWithDefault<int>(0) == 1);
-    }
-
-    SECTION("text input focused: bare-key suppressed") {
-        TextEditor input;
-        input.claim_input_focus();
-        REQUIRE(View::focused_input_ == &input);
-        REQUIRE(input.accepts_text_input());
-
-        bridge.forward_key_event(63, 0, true);
-        REQUIRE(engine.evaluate("bare_count").getWithDefault<int>(0) == 0);
-
-        // Shift alone is also guarded.
-        bridge.forward_key_event(63, 1, true);
-        REQUIRE(engine.evaluate("shift_count").getWithDefault<int>(0) == 0);
-
-        // Modifier chord still fires — Cmd+, opens Settings even from an
-        // input.
-        bridge.forward_key_event(44, 16, true);
-        REQUIRE(engine.evaluate("cmd_count").getWithDefault<int>(0) == 1);
-
-        input.release_input_focus();
-        REQUIRE(View::focused_input_ == nullptr);
-    }
-
-    SECTION("non-text focusable (knob/button) focused: bare-key still fires") {
-        // Codex review pin (#2120): `focused_input_` is claimed by any
-        // focusable widget, not just text inputs. The guard MUST check
-        // `accepts_text_input()` — otherwise clicking a knob would kill
-        // every global single-key shortcut until focus moved away.
-        View knob_like;
-        knob_like.set_focusable(true);
-        knob_like.claim_input_focus();
-        REQUIRE(View::focused_input_ == &knob_like);
-        REQUIRE_FALSE(knob_like.accepts_text_input());
-
-        bridge.forward_key_event(63, 0, true);
-        REQUIRE(engine.evaluate("bare_count").getWithDefault<int>(0) == 1);
-
-        bridge.forward_key_event(63, 1, true);
-        REQUIRE(engine.evaluate("shift_count").getWithDefault<int>(0) == 1);
-
-        knob_like.release_input_focus();
-    }
-
-    SECTION("focus released: bare-key fires again") {
-        TextEditor input;
-        input.claim_input_focus();
-        bridge.forward_key_event(63, 0, true);
-        REQUIRE(engine.evaluate("bare_count").getWithDefault<int>(0) == 0);
-
-        input.release_input_focus();
-        bridge.forward_key_event(63, 0, true);
-        REQUIRE(engine.evaluate("bare_count").getWithDefault<int>(0) == 1);
-    }
-}
